Server/requirements.txt
flask
flask_login
flask_wtf
flask_table
flask_bootstrap
redis
pillow